Move all macros used in aarch64.asm into asm_offsets.h.
It was also necessary to move some declarations around.
