+++ core/arch/aarch64/asm_offsets.cpp
@@ -31,6 +31,8 @@
  */
 
 #include "../globals.h"
+#include "../../unix/module_private.h"
+#include "mangle.h"
 #include <stddef.h> /* for offsetof */
 #include "asm_offsets.h"
 
@@ -41,5 +43,26 @@ namespace dynamorio {
 #define CHECK(x) static_assert(x, "macro in asm_offsets.h defined incorrectly")
 
 CHECK(dcontext_t_OFFSET_dstack == offsetof(dcontext_t, dstack));
+CHECK(dcontext_t_OFFSET_is_exiting == offsetof(dcontext_t, is_exiting));
+
+CHECK(icache_op_struct_t_OFFSET_flag == offsetof(icache_op_struct_t, flag));
+CHECK(icache_op_struct_t_OFFSET_lock == offsetof(icache_op_struct_t, lock));
+CHECK(icache_op_struct_t_OFFSET_linesize == offsetof(icache_op_struct_t, linesize));
+CHECK(icache_op_struct_t_OFFSET_begin == offsetof(icache_op_struct_t, begin));
+CHECK(icache_op_struct_t_OFFSET_end == offsetof(icache_op_struct_t, end));
+CHECK(icache_op_struct_t_OFFSET_spill == offsetof(icache_op_struct_t, spill));
+
+CHECK(priv_mcontext_t_OFFSET_simd == offsetof(priv_mcontext_t, simd));
+CHECK(priv_mcontext_t_SIZE == sizeof(priv_mcontext_t));
+
+CHECK(spill_state_t_OFFSET_r0 == offsetof(spill_state_t, r0));
+CHECK(spill_state_t_OFFSET_r1 == offsetof(spill_state_t, r1));
+CHECK(spill_state_t_OFFSET_r2 == offsetof(spill_state_t, r2));
+CHECK(spill_state_t_OFFSET_r3 == offsetof(spill_state_t, r3));
+CHECK(spill_state_t_OFFSET_r4 == offsetof(spill_state_t, r4));
+CHECK(spill_state_t_OFFSET_r5 == offsetof(spill_state_t, r5));
+CHECK(spill_state_t_OFFSET_fcache_return == offsetof(spill_state_t, fcache_return));
+
+CHECK(struct_tlsdesc_t_OFFSET_arg == offsetof(struct tlsdesc_t, arg));
 
 } // namespace dynamorio
